[ADD] 좋아요 기능 추가, 게시글list 조회 추가 - 데이터 구동 확인 필요

+++ back/src/routers/commentRouter.js
@@ -1,21 +1,24 @@
 import express from "express";
-import Comments from "../../db/models/comments";
-import Users from "../../db/models/user";
 import mysqlManager from "../../db";
 import Sequelize from "sequelize";
+import { Comments, Users, Postings } from "../../db/models/comments";
 
 const commentRouter = express.Router();
 
 // 댓글 생성
 commentRouter.post("/:postings_id/comment", async (req, res, next) => {
   try {
-    const comments = {
+    const posting = await Postings.findOne({ where: { id: req.params.postings_id } });
+    if (!posting) {
+      return res.status(403).send("존재하지 않는 게시글입니다.");
+    }
+    const comment = await Comment.create({
       users_id: req.body.users_id,
       postings_id: req.params.postings_id,
       content: req.body.content,
-    };
+    });
     const fullComment = Comment.findOne({
-      where: { id: comments.id },
+      where: { id: comment.id },
       include: [
         {
           model: Users,
@@ -30,32 +33,17 @@ commentRouter.post("/:postings_id/comment", async (req, res, next) => {
   }
 });
 
-/** ## 게시글 조회할 때, 그 게시글의 댓글들 모두 불러옴으로 이 router는 필요없다고 판단
-
-// 특정 게시글의 모든 댓글 조회 -
-commentRouter.get("/postings/:postings_id/comments", async (req, res, next) => {
-  try {
-    const postings_id = req.params.id;
-    Comments.findAll({
-      where: { id: postings_id },
-    }).then((result) => {
-      res.status(200).json(result);
-    });
-  } catch (error) {
-    next(error);
-  }
-});
-*/
-
 // 댓글 수정
 commentRouter.put("/:postings_id/comments/:id", async (req, res, next) => {
   try {
-    const id = req.params.id;
-    const postings_id = req.params.id;
-
-    await Comments.update({ content: req.body.content }, { where: { id } });
+    const posting = await Postings.findOne({ where: { id: req.params.postings_id } });
+    if (!posting) {
+      return res.status(403).send("존재하지 않는 게시글입니다.");
+    }
+    const comment = await Comments.findOne({ where: { id: req.params.id } });
+    await comment.update({ content: req.body.content }, { where: { id: req.params.id } });
     const updatedComment = await Comments.findOne({
-      where: { id },
+      where: { id: req.params.id },
       include: [
         {
           model: Users,
@@ -70,11 +58,14 @@ commentRouter.put("/:postings_id/comments/:id", async (req, res, next) => {
 });
 
 // 댓글 삭제
-commentRouter.delete("/:postings_id/comments/:id", (req, res, next) => {
+commentRouter.delete("/:postings_id/comments/:id", async (req, res, next) => {
   try {
-    const id = req.params.id;
+    const posting = await Postings.findOne({ where: { id: req.params.postings_id } });
+    if (!posting) {
+      return res.status(403).send("존재하지 않는 게시글입니다.");
+    }
     Comments.destroy({
-      where: { id },
+      where: { id: req.params.id },
     });
     res.json({
       postings_id: req.params.postings_id,
